Show ErrorHandlerView on proper server error, update spring logo
